[core] Add bulk load for bootstraping in ClusteringCompactManager
